refactor: refact ts to js
