+++ components/shared/Empty.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export const Empty = () => {
+    return (
+      <div className="empty">
+        <div className="empty-icon h1">🙀</div>
+        <p className="empty-title h4">No results found</p>
+        <p className="empty-subtitle">Please try different identity keyword.</p>
+      </div>
+    )
+  };
