+++ components/shared/Loading.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const Loading = () => {
+  return (
+    <div className="loading-container">
+      <div className="loading"></div>
+    </div>
+  );
+};
