+++ package.json
@@ -20,6 +20,7 @@
     "sass": "^1.53.0"
   },
   "devDependencies": {
+    "@types/node": "^18.7.14",
     "eslint": "8.19.0",
     "eslint-config-next": "12.2.2"
   }
