+++ pages/_app.tsx
@@ -1,9 +1,10 @@
+import React from "react";
 import { ApolloProvider } from "@apollo/client";
 import client from "../utils/apollo";
 import { GoogleAnalytics } from "nextjs-google-analytics";
 import "../styles/scss/web3bio.scss";
 
-function App({ Component, pageProps }) {
+export default function App({ Component, pageProps }) {
   return (
     <ApolloProvider client={client}>
       <Component {...pageProps} />
@@ -11,5 +12,3 @@ function App({ Component, pageProps }) {
     </ApolloProvider>
   );
 }
-
-export default App;
